Began reworking skeleton export to secure correct associations.
Exporting skeletons solely from an ISkin could mean that any unbound joints are
excluded. It also leaves the joint order to the internals of MAX. Both of these
circumstances could result in incorrect ordering meaning that the indices in the
JOINT_INDEX stream do not correctly match the actual indices of joints.

The new method is to export a skeleton by traversing the nodes, and to store it
in an intermediate structure to facilitate look-up of indices during geometry
export.